+++ packages/server/src/server/agent/agent-mcp.e2e.test.ts
@@ -89,6 +89,63 @@ async function createMcpClient(url: string, authToken?: string): Promise<McpClie
   return { callTool: boundCallTool, close: () => rawClient.close() };
 }
 
+interface LaunchRecorder {
+  recordedLaunches: AgentSessionConfig[];
+}
+
+class RecordingAgentClient implements AgentClient {
+  readonly provider: AgentClient["provider"];
+  readonly capabilities: AgentClient["capabilities"];
+
+  constructor(
+    private readonly inner: AgentClient,
+    private readonly recorder: LaunchRecorder,
+  ) {
+    this.provider = inner.provider;
+    this.capabilities = {
+      ...inner.capabilities,
+      supportsMcpServers: true,
+      supportsNativePaseoTools: false,
+    };
+  }
+
+  async createSession(
+    ...args: Parameters<AgentClient["createSession"]>
+  ): ReturnType<AgentClient["createSession"]> {
+    this.recorder.recordedLaunches.push(args[0]);
+    return this.inner.createSession(...args);
+  }
+
+  async resumeSession(
+    ...args: Parameters<AgentClient["resumeSession"]>
+  ): ReturnType<AgentClient["resumeSession"]> {
+    return this.inner.resumeSession(...args);
+  }
+
+  async fetchCatalog(
+    ...args: Parameters<AgentClient["fetchCatalog"]>
+  ): ReturnType<AgentClient["fetchCatalog"]> {
+    return this.inner.fetchCatalog(...args);
+  }
+
+  async isAvailable(): Promise<boolean> {
+    return this.inner.isAvailable();
+  }
+}
+
+function createMcpRecordingAgentClients(recorder: LaunchRecorder) {
+  const clients = createTestAgentClients();
+  const claude = clients.claude;
+  if (!claude) {
+    throw new Error("Fake Claude client is not configured");
+  }
+
+  return {
+    ...clients,
+    claude: new RecordingAgentClient(claude, recorder),
+  };
+}
+
 async function waitForAgentCompletion(options: {
   client: McpClient;
   agentId: string;
@@ -256,6 +313,7 @@ describe("agent MCP end-to-end (offline)", () => {
     const staticDir = await mkdtemp(path.join(os.tmpdir(), "paseo-static-"));
     const agentCwd = await mkdtemp(path.join(os.tmpdir(), "paseo-agent-cwd-"));
     const port = await getAvailablePort();
+    const recorder: LaunchRecorder = { recordedLaunches: [] };
 
     const daemonConfig: PaseoDaemonConfig = {
       listen: `127.0.0.1:${port}`,
@@ -265,7 +323,7 @@ describe("agent MCP end-to-end (offline)", () => {
       mcpEnabled: true,
       staticDir,
       mcpDebug: false,
-      agentClients: createTestAgentClients(),
+      agentClients: createMcpRecordingAgentClients(recorder),
       agentStoragePath: path.join(paseoHome, "agents"),
     };
 
@@ -278,6 +336,7 @@ describe("agent MCP end-to-end (offline)", () => {
     const disabledStaticDir = await mkdtemp(path.join(os.tmpdir(), "paseo-static-disabled-"));
     const disabledAgentCwd = await mkdtemp(path.join(os.tmpdir(), "paseo-agent-cwd-disabled-"));
     const disabledPort = await getAvailablePort();
+    const disabledRecorder: LaunchRecorder = { recordedLaunches: [] };
     const disabledDaemonConfig: PaseoDaemonConfig = {
       listen: `127.0.0.1:${disabledPort}`,
       paseoHome: disabledPaseoHome,
@@ -287,7 +346,7 @@ describe("agent MCP end-to-end (offline)", () => {
       mcpInjectIntoAgents: false,
       staticDir: disabledStaticDir,
       mcpDebug: false,
-      agentClients: createTestAgentClients(),
+      agentClients: createMcpRecordingAgentClients(disabledRecorder),
       agentStoragePath: path.join(disabledPaseoHome, "agents"),
     };
     const disabledDaemon = await createPaseoDaemon(disabledDaemonConfig, pino({ level: "silent" }));
@@ -313,13 +372,14 @@ describe("agent MCP end-to-end (offline)", () => {
       agentId = typeof payload?.agentId === "string" ? payload.agentId : null;
       expect(agentId).toBeTruthy();
 
-      const injectedAgent = daemon.agentManager.getAgent(agentId!);
-      expect(injectedAgent?.config.mcpServers).toMatchObject({
+      expect(recorder.recordedLaunches.at(-1)?.mcpServers).toMatchObject({
         paseo: {
           type: "http",
           url: `http://127.0.0.1:${port}/mcp/agents?callerAgentId=${agentId!}`,
         },
       });
+      const injectedAgent = daemon.agentManager.getAgent(agentId!);
+      expect(injectedAgent?.config.mcpServers?.paseo).toBeUndefined();
 
       const disabledResult = await disabledClient.callTool({
         name: "create_agent",
@@ -337,6 +397,7 @@ describe("agent MCP end-to-end (offline)", () => {
         typeof disabledPayload?.agentId === "string" ? disabledPayload.agentId : null;
       expect(disabledAgentId).toBeTruthy();
 
+      expect(disabledRecorder.recordedLaunches.at(-1)?.mcpServers?.paseo).toBeUndefined();
       const disabledAgent = disabledDaemon.agentManager.getAgent(disabledAgentId!);
       expect(disabledAgent?.config.mcpServers?.paseo).toBeUndefined();
     } finally {
@@ -364,6 +425,7 @@ describe("agent MCP end-to-end (offline)", () => {
     const staticDir = await mkdtemp(path.join(os.tmpdir(), "paseo-static-"));
     const agentCwd = await mkdtemp(path.join(os.tmpdir(), "paseo-agent-cwd-"));
     const port = await getAvailablePort();
+    const recorder: LaunchRecorder = { recordedLaunches: [] };
 
     const daemonConfig: PaseoDaemonConfig = {
       listen: `0.0.0.0:${port}`,
@@ -373,7 +435,7 @@ describe("agent MCP end-to-end (offline)", () => {
       mcpEnabled: true,
       staticDir,
       mcpDebug: false,
-      agentClients: createTestAgentClients(),
+      agentClients: createMcpRecordingAgentClients(recorder),
       agentStoragePath: path.join(paseoHome, "agents"),
     };
 
@@ -399,13 +461,14 @@ describe("agent MCP end-to-end (offline)", () => {
       agentId = typeof payload?.agentId === "string" ? payload.agentId : null;
       expect(agentId).toBeTruthy();
 
-      const injectedAgent = daemon.agentManager.getAgent(agentId!);
-      expect(injectedAgent?.config.mcpServers).toMatchObject({
+      expect(recorder.recordedLaunches.at(-1)?.mcpServers).toMatchObject({
         paseo: {
           type: "http",
           url: `http://127.0.0.1:${port}/mcp/agents?callerAgentId=${agentId!}`,
         },
       });
+      const injectedAgent = daemon.agentManager.getAgent(agentId!);
+      expect(injectedAgent?.config.mcpServers?.paseo).toBeUndefined();
     } finally {
       if (agentId) {
         await client.callTool({ name: "kill_agent", args: { agentId } });
@@ -568,6 +631,23 @@ describe("agent MCP end-to-end (offline)", () => {
         return true;
       }
 
+      async fetchCatalog(): Promise<{
+        models: Array<{ provider: "codex"; id: string; label: string; isDefault: boolean }>;
+        modes: Array<{ id: string; label: string; description: string }>;
+      }> {
+        return {
+          models: [
+            {
+              provider: "codex",
+              id: "gpt-5.4-mini",
+              label: "gpt-5.4-mini",
+              isDefault: true,
+            },
+          ],
+          modes: [{ id: "full-access", label: "Full access", description: "No prompts" }],
+        };
+      }
+
       async createSession(_config: AgentSessionConfig): Promise<AgentSession> {
         return new StartTurnFailureSession();
       }
@@ -605,6 +685,7 @@ describe("agent MCP end-to-end (offline)", () => {
 
     const client = await createMcpClient(`http://127.0.0.1:${port}/mcp/agents`);
 
+    let agentId: string | null = null;
     try {
       const result = await client.callTool({
         name: "create_agent",
@@ -618,14 +699,25 @@ describe("agent MCP end-to-end (offline)", () => {
         },
       });
 
-      expect(result.isError).toBe(true);
-      const contentItem = result.content?.[0];
-      const contentText: string | undefined =
-        contentItem != null && typeof contentItem === "object"
-          ? Reflect.get(contentItem, "text")
-          : undefined;
-      expect(contentText ?? "").toContain("Initial turn failed to start");
+      const payload = getStructuredContent(result);
+      agentId = typeof payload?.agentId === "string" ? payload.agentId : null;
+      expect(agentId).toBeTruthy();
+
+      await waitForAgentCompletion({ client, agentId: agentId! });
+      const statusResult = await client.callTool({
+        name: "get_agent_status",
+        args: { agentId },
+      });
+      const statusPayload = getStructuredContent(statusResult);
+      expect(statusPayload?.status).toBe("error");
+      const snapshot = statusPayload?.snapshot;
+      const lastError =
+        snapshot && typeof snapshot === "object" ? Reflect.get(snapshot, "lastError") : undefined;
+      expect(lastError).toContain("Initial turn failed to start");
     } finally {
+      if (agentId) {
+        await client.callTool({ name: "kill_agent", args: { agentId } });
+      }
       await client.close();
       await daemon.stop();
       await rm(paseoHome, { recursive: true, force: true });
