+++ packages/server/src/server/agent/tools/types.ts
@@ -1,0 +1,44 @@
+import type { z } from "zod";
+
+export interface PaseoToolExecutionContext {
+  signal?: AbortSignal;
+}
+
+export interface PaseoToolResult {
+  content: Array<{ type: string; text?: string; [key: string]: unknown }>;
+  structuredContent?: unknown;
+  isError?: boolean;
+}
+
+export interface PaseoToolConfig {
+  title?: string;
+  description?: string;
+  inputSchema?: z.ZodRawShape | z.ZodType;
+  outputSchema?: z.ZodRawShape | z.ZodType;
+}
+
+export interface PaseoToolDefinition extends PaseoToolConfig {
+  name: string;
+  description: string;
+  handler: (input: unknown, context: PaseoToolExecutionContext) => Promise<PaseoToolResult>;
+}
+
+export interface PaseoToolCatalog {
+  tools: ReadonlyMap<string, PaseoToolDefinition>;
+  getTool(name: string): PaseoToolDefinition | undefined;
+  executeTool(
+    name: string,
+    input: unknown,
+    context?: PaseoToolExecutionContext,
+  ): Promise<PaseoToolResult>;
+}
+
+export interface PaseoToolRuntimeContext {
+  callerAgentId?: string;
+  enableVoiceTools?: boolean;
+  voiceOnly?: boolean;
+}
+
+export type PaseoToolCatalogFactory = (
+  context: PaseoToolRuntimeContext,
+) => PaseoToolCatalog | Promise<PaseoToolCatalog>;
